requirements/base.txt
celery==5.3.0
django==4.2.4
django-allauth==0.55.0
django-filter==23.2
psycopg2-binary==2.9.7
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-modeltranslation==0.18.11
django-elasticsearch-dsl==7.3
django-mptt==0.14.0
PaycomUz==3.0
Pillow==10.0.0
pandas==2.0.3
djangorestframework-simplejwt==5.3.0
django-embed-video==1.4.9
drf-yasg==1.21.7
dj-rest-auth==4.0.1
facebook-sdk==3.1.0
google-api-core==2.11.1
google-api-python-client==2.96.0
redis==4.6.0
pyTelegramBotAPI==4.12.0
django_import-export==3.2.0
django-ckeditor==6.7.0
opencv-python==4.8.0.76
stripe==6.0.0
pytube==15.0.0
